fix eventduble and singlecand tarfile
git-svn-id: http://svn.cern.ch/guest/lhcb/Analysis/trunk@132411 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 1e461997d807bbf16d8989c4c3afecbe44c4192c

+++ Analysis/Phys/FlavourTaggingChecker/src/BTaggingAnalysis.cpp
@@ -148,7 +148,8 @@ StatusCode BTaggingAnalysis::execute() {
   debug()<<">>>>>  Processing Run "<<evt->runNumber() 
          <<"   Event "<<evt->evtNumber()<<"  <<<<<" <<endreq;
   tuple -> column ("run", evt->runNumber());
-  tuple -> column ("event", (long)evt->evtNumber());
+  //  tuple -> column ("event", (long)evt->evtNumber());
+  tuple -> column ("event", (double)evt->evtNumber());
 
   //----------------------------------------------------------------------
   //MCinfo
